fix: updating values for relations nested in lists
There is a bug where modifying results that are nested within a list
result does not work, the result is always undefined.

The reason modifying results that are within a list does not work is
that the nested relation is found on each item of the list rather than
the list itself, so when we get the relation by name the result is
always `undefined`.

When a `next` function is returning a relation that is nested within a
list combine all the relation values into a single flat array. This
means middleware only has to handle flat arrays of results which makes
modifying the result before it is returned easier. If the result's
parent is needed then it is possible to go through the parent middleware
and traverse that relation; only a single depth of relation needs to be
traversed as the middleware will be called for each layer.

Add more specific tests to ensure the problem is fixed at various
depths.