fix: returning null relations in nested arrays
When result returned from the Prisma Client has relations with null
values nested in parent arrays the `setNestedResult` helper mistakenly
sets those values to an empty object.

Don't include null values in the results for the relevant nested
middleware call, and explicitly set the relations result to null when
they are not found in the results array.