feat: ensure nested create args match create args
When calling the middleware function for create actions the structure of
the args object depends on whether it was a nested create or not. When
the create is the root operation the data is found in the data field of
args but for nested creates the args are the data.

To make it simpler to write middleware for nested creates change the
args in params to match root creates: put the data in a data field.

Any modifications made to the args object outside the data object are
ignored for nested creates.

BREAKING CHANGE: the data in nested creates is now nested in a data
field, middleware that handled nested creates will be broken.