feat: call middleware for writes in nested lists
In a prisma query it is possible to provide multiple create, update or
delete operations within a query as a list. To avoid nested middleware
needing to handle these nested lists manually call the middleware
function for each write within the list. This means that nested lists of
write operations can be handled in the same way as when they are not in
a list.

This is not a breaking change due to the fact that middleware would need
to handle the non-list case along with any lists. The new behaviour will
cause list operations to be handled by the non-list logic.