feat: call middleware for all lists of operations
Similarly to nested update, create or delete operations it is possible
to pass an array of operations to nested upsert, updateMany and
deleteMany actions. Since createMany operations can be flattened to a
single data array they cannot be passed as an array.

For update, create and delete actions we already call the passed
middleware function for each operation in the array if passed,
unfortunately other actions that support passing an array of operations
were omitted.

On closer inspection the Prisma nested write API is such that the only
time an action is passed as an array is when it is an operation list.
Due to this it's possible to change the logic to simply check for
whether the action is an array in order to determine if the middleware
function should be called for each operation in the array.

This is not a breaking change due to the fact that middleware already
needs to handle the non-list case along with any lists. The new
behaviour will cause list operations to be handled by the non-list
logic.