feat: include relation in NestedParams
For certain middleware it is necessary to reference the relation for
that nested action. For example instead of deleting a toOne relation a
soft delete middleware may decide instead to unlink the relation and
update it to have a deleted mark which would require both the isList and
the foreign key information stored on the relation.