fix: queries that use Json NullTypes
Prisma NullTypes are used as the null value for Json fields. When using
a NullType in a query the instance of the NullType must remain the same
or Prisma is unable to parse the value correctly, replacing it with
undefined.

Lodash cloneDeep causes these NullTypes to be new instances and so
params that use them have unexpected undefined values, causing errors.

When the Prisma client has NullTypes, which is true for v4, clone params
but pass through the instances of Prisma.JsonNull, Prisma.DbNull and
Prisma.AnyNull so that they are not replaced with undefined.

Prisma client v3 uses strings for these null types so cloneDeep works as
is in that case.