fix: error parsing invalid logical where arrays (#22)
When extracting nested actions in logical where arrays we can encounter
an error if the array contains null or undefined. This is because we try
to access a property on those values.

Use optional chaining to ensure that the value is defined before trying
to access a property in logical where arrays.