feat: avoid multiple types of select calls
Currently there are two scenarios when a select action type is found:
- When an include has a select within it
- When a select has a relation in it

This means that there is a situation where it is not possible to reason
about what to do with a select because the two scenarios look the same
based on the available information; both have the include action as the
parent.

BREAKING CHANGE: remove the calls for include selects

Middleware that relied on calls with the "select" action for select
objects within an include will no longer be able to use that call.
Instead use the parent "include" action to modify the selected fields,
or use the "select" action for the relation within that select object.